+++ src/main.rs
@@ -1,10 +1,15 @@
 #![forbid(unsafe_code)]
 #![deny(clippy::all)]
 
+use anyhow::Context;
 use slumber_cli::Args;
-use slumber_core::util::{DataDirectory, TempDirectory};
+use slumber_core::util::{DataDirectory, ResultTraced};
 use slumber_tui::Tui;
-use std::{fs::File, io, process::ExitCode};
+use std::{
+    fs::{self, File, OpenOptions},
+    io,
+    process::ExitCode,
+};
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::{filter::Targets, fmt::format::FmtSpan, prelude::*};
 
@@ -13,12 +18,8 @@ async fn main() -> anyhow::Result<ExitCode> {
     // Global initialization
     let args = Args::parse();
     DataDirectory::init()?;
-    TempDirectory::init()?;
-    initialize_tracing(args.subcommand.is_some()).unwrap();
 
-    if args.global.log {
-        println!("{}", TempDirectory::get().log().display());
-    }
+    initialize_tracing(args.subcommand.is_some());
 
     // Select mode based on whether request ID(s) were given
     match args.subcommand {
@@ -46,25 +47,35 @@ async fn main() -> anyhow::Result<ExitCode> {
     }
 }
 
-/// Set up tracing to log to a file. Optionally also log to stderr (for CLI
-/// usage)
-fn initialize_tracing(console_output: bool) -> anyhow::Result<()> {
-    let path = TempDirectory::get().log();
-    let log_file = File::create(path)?;
+/// Set up tracing to a log file, and optionally the console as well. If there's
+/// an error creating the log file, we'll skip that part. This means in the TUI
+/// the error (and all other tracing) will never be visible, but that's a
+/// problem for another day.
+fn initialize_tracing(console_output: bool) {
+    // Failing to log shouldn't be a fatal crash, so just move on
+    let log_file = initialize_log_file()
+        .context("Error creating log file")
+        .traced()
+        .ok();
+
     // Basically a minimal version of EnvFilter that doesn't require regexes
     // https://github.com/tokio-rs/tracing/issues/1436#issuecomment-918528013
     let targets: Targets = std::env::var("RUST_LOG")
         .ok()
         .and_then(|env| env.parse().ok())
         .unwrap_or_default();
-    let file_subscriber = tracing_subscriber::fmt::layer()
-        .with_file(true)
-        .with_line_number(true)
-        .with_writer(log_file)
-        .with_target(false)
-        .with_ansi(false)
-        .with_span_events(FmtSpan::NEW)
-        .with_filter(targets);
+    let file_subscriber = log_file.map(|log_file| {
+        // Include PID
+        // https://github.com/tokio-rs/tracing/pull/2655
+        tracing_subscriber::fmt::layer()
+            .with_file(true)
+            .with_line_number(true)
+            .with_writer(log_file)
+            .with_target(false)
+            .with_ansi(false)
+            .with_span_events(FmtSpan::NEW)
+            .with_filter(targets)
+    });
 
     // Enable console output for CLI
     let console_subscriber = if console_output {
@@ -84,5 +95,26 @@ fn initialize_tracing(console_output: bool) -> anyhow::Result<()> {
         .with(file_subscriber)
         .with(console_subscriber)
         .init();
-    Ok(())
+}
+
+/// Create the log file. If it already exists, make sure it's not over a max
+/// size. If it is, move it to a backup path and nuke whatever might be in the
+/// backup path.
+fn initialize_log_file() -> anyhow::Result<File> {
+    const MAX_FILE_SIZE: u64 = 1000 * 1000; // 1MB
+    let data_directory = DataDirectory::get();
+    let path = data_directory.log_file();
+
+    if fs::metadata(&path)
+        .map_or(false, |metadata| metadata.len() > MAX_FILE_SIZE)
+    {
+        // Rename new->old, overwriting old. If that fails, just delete new so
+        // it doesn't grow indefinitely. Failure shouldn't stop us from logging
+        // though
+        let _ = fs::rename(&path, data_directory.log_file_old())
+            .or_else(|_| fs::remove_file(&path));
+    }
+
+    let log_file = OpenOptions::new().create(true).append(true).open(path)?;
+    Ok(log_file)
 }